backend/pyproject.toml
[project]
name = "newpizza"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel Bushmelev",email = "96923234+Dikroll@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "asgiref (==3.8.1)",
    "django (==5.1.1)",
    "django-cors-headers (==4.4.0)",
    "djangorestframework (==3.15.2)",
    "djangorestframework-simplejwt (==5.4.0)",
    "django-debug-toolbar (==5.0.1)",
    "pillow (==10.4.0)",
    "sqlparse (==0.5.1)",
    "tzdata (==2024.2)",
    "psycopg2-binary (==2.9.10)",
    "python-decouple (==3.8)",
    "django-redis (==5.4.0)",
    "python-dotenv (==1.0.1)",
    "gunicorn (==22.0.0)",
    "requests (==2.26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
